requirements.txt
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
Flask==3.1.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
git-filter-repo==2.47.0
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4

